Rework handling of value groups
Single values are far more common than value groups, so it makes more
sense to model the API around single values by default, and represent
value groups as a special kind of single value. This simplifies many
uses of the API, because it is no longer necessary to unwrap
single-value "groups".