Don't resolve object references automatically during reading
Instead they are now returned as ObjectReference objects, which can be
resolved afterwards in a separate step. This makes the read data
structure less convenient to use at the moment (because there is no
built-in way to resolve the references yet), but allows safely reading
and displaying data with circular references.