Rework how _init_from_unarchiver_ implementations are called
All _init_from_unarchiver_ implementations in a class chain are now
called automatically - the individual implementations are no longer
responsible for calling their superclass's implementation. The checks
for the archived class name and superclass are now also performed
automatically before _init_from_unarchiver_ is called, which eliminates
almost all of the manual boilerplate code in the implementations (the
only remaining manual checks are for the version number).

Unfortunately this new implementation is relatively unconventional and
complex - it has to work around Python's normal inheritance mechanisms
and instead walk the superclass chain and inspect __dict__ manually, so
that it can control when and how the _init_from_unarchiver_
implementation from each class is called, and modify arguments and add
the appropriate checks in between. At least it's simpler for the user
code in the end, hopefully.