Add initial support for unarchiving structs to custom Python classes
This is similar to the mechanism for objects, but the Python logic is
much simpler, because structs don't support inheritance or references.
Some extra logic is still needed though, because sometimes the same
struct type can be encoded either normally or with its name replaced
with "?". This complicates both the expected encoding checks and the
custom struct class mapping a bit.