Use contextvars to manage RecursiveReprState instead of explicit args
Literally the only code that actually uses the state is the base
implementation of _as_multiline_string_. All other multiline string
method implementations receive the state and sometimes pass it on to
other calls, but otherwise don't really use it. Managing the state using
a ContextVar simplifies the implementations a lot.